Rename SSO integration e2e tests file
